Update user profile update / get event people

+++ module/Application/src/Application/memreas/AddMediaEvent.php
@@ -183,7 +183,7 @@ error_log ( "location ---> " . $location . PHP_EOL );
 
                 if ($is_profile_pic) {
                     //Remove previous profile images
-                    $remove_old_profile = "DELETE FROM Application\Entity\Media m WHERE m.is_profile_pic = 1 AND m.media_id <> '$media_id'";
+                    $remove_old_profile = "DELETE FROM Application\Entity\Media m WHERE m.is_profile_pic = 1 AND m.media_id <> '$media_id' AND m.user_id = '$user_id'";
                     $remove_result = $this->dbAdapter->createQuery($remove_old_profile);
                     $remove_result->getResult();
